+++ v3/crates/compatibility/src/compatibility_date.rs
@@ -163,7 +163,7 @@ pub fn get_compatibility_date_for_flag(flag: Flag) -> Option<CompatibilityDate> 
         | Flag::DisallowLocalRelationshipsOnDataConnectorsWithoutRelationshipsOrVariables => {
             Some(new_compatibility_date(2025, 2, 8))
         }
-        Flag::DisallowRecursiveObjectTypes => {
+        Flag::DisallowUnknownValuesInArguments | Flag::DisallowRecursiveObjectTypes => {
             // TODO Set a compatibility date during the next release
             None
         }